+++ src/arkiver/types.ts
@@ -114,3 +114,50 @@ export type EventHandler<
 > = (ctx: EventHandlerContext<TAbiEvent, TEventName>) => Promise<void>;
 
 export type BlockHandler = (ctx: BlockHandlerContext) => Promise<void>;
+
+type ValidNameChars =
+  | "a"
+  | "b"
+  | "c"
+  | "d"
+  | "e"
+  | "f"
+  | "g"
+  | "h"
+  | "i"
+  | "j"
+  | "k"
+  | "l"
+  | "m"
+  | "n"
+  | "o"
+  | "p"
+  | "q"
+  | "r"
+  | "s"
+  | "t"
+  | "u"
+  | "v"
+  | "w"
+  | "x"
+  | "y"
+  | "z"
+  | "0"
+  | "1"
+  | "2"
+  | "3"
+  | "4"
+  | "5"
+  | "6"
+  | "7"
+  | "8"
+  | "9"
+  | "-"
+  | "_";
+
+export type CheckManifestName<Name extends string, FullName extends string> =
+  Name extends `${infer First}${infer Rest}`
+    ? First extends ValidNameChars | Capitalize<ValidNameChars>
+      ? CheckManifestName<Rest, FullName>
+    : `Invalid character in manifest name: ${First}`
+    : FullName;